requirements.txt
flask
openai-whisper
librosa
soundfile
torch